requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0
PyMySQL==1.1.0
cryptography==46.0.2
